rename QueryCondition interface and cleanup
