+++ src/main/java/com/teragrep/pth_06/planner/walker/conditions/QueryCondition.java
@@ -47,121 +47,7 @@
 package com.teragrep.pth_06.planner.walker.conditions;
 
 import org.jooq.Condition;
-import org.jooq.Table;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.w3c.dom.Element;
 
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-
-/**
- * Creates a query condition from provided dom element
- */
-public final class QueryCondition {
-    private static final Logger LOGGER = LoggerFactory.getLogger(QueryCondition.class);
-
-    private final Element element;
-    private final QueryConditionConfig config;
-    private final Set<Table<?>> tableSet;
-    private final long bloomTermId;
-    private final List<Condition> conditionCache;
-
-    private void checkValidity(Element element) {
-        if (element.getTagName() == null) {
-            throw new IllegalStateException("Tag name for Element was null");
-        }
-        if (!element.hasAttribute("operation")) {
-            throw new IllegalStateException("Could not find specified or default value for 'operation' attribute from Element");
-        }
-        if (!element.hasAttribute("value")) {
-            throw new IllegalStateException("Could not find specified or default value for 'value' attribute from Element");
-        }
-    }
-
-    public QueryCondition(Element element, QueryConditionConfig config, long bloomTermId) {
-        this.element = element;
-        this.config = config;
-        this.bloomTermId = bloomTermId;
-        this.tableSet = new HashSet<>();
-        this.conditionCache = new ArrayList<>(1);
-    }
-
-    public Condition condition() {
-        generateCondition();
-        Condition condition = conditionCache.get(0);
-        LOGGER.debug("Query condition: <{}>", condition);
-        return condition;
-    }
-
-    public void generateCondition() {
-        if (!conditionCache.isEmpty()) {
-            return;
-        }
-        checkValidity(element);
-        String tag = element.getTagName();
-        Condition condition = null;
-        switch (tag.toLowerCase()) {
-            case "index":
-                Conditionable index = new IndexCondition(element, config.streamQuery());
-                condition = index.condition();
-                break;
-            case "sourcetype":
-                Conditionable sourceType = new SourceTypeCondition(element, config.streamQuery());
-                condition = sourceType.condition();
-                break;
-            case "host":
-                Conditionable host = new HostCondition(element, config.streamQuery());
-                condition = host.condition();
-                break;
-        }
-        if (!config.streamQuery()) {
-            // Handle also time qualifiers
-            if ("earliest".equalsIgnoreCase(tag) || "index_earliest".equalsIgnoreCase(tag)) {
-                Conditionable earliest = new EarliestCondition(element);
-                condition = earliest.condition();
-            }
-            if ("latest".equalsIgnoreCase(tag) || "index_latest".equalsIgnoreCase(tag)) {
-                Conditionable latest = new LatestCondition(element);
-                condition = latest.condition();
-            }
-            // value search
-            if ("indexstatement".equalsIgnoreCase(tag) && config.bloomEnabled()) {
-                IndexStatementCondition indexStatementCondition =
-                        new IndexStatementCondition(condition, element, config, bloomTermId);
-                condition = indexStatementCondition.condition();
-                List<Table<?>> newMatches = indexStatementCondition.matchList();
-                if (!newMatches.isEmpty()) {
-                    tableSet.addAll(newMatches);
-                }
-            }
-        }
-        conditionCache.add(condition);
-    }
-
-    public boolean isIndexStatement() {
-        checkValidity(element);
-        String tag = element.getTagName();
-        return !config.streamQuery() && "indexstatement".equalsIgnoreCase(tag) && config.bloomEnabled();
-    }
-
-    public Set<Table<?>> matchList() {
-        checkValidity(element);
-        generateCondition();
-        return tableSet;
-    }
-
-    public long earliest() {
-        checkValidity(element);
-        String tag = element.getTagName();
-        int earliestFromElement = Integer.MAX_VALUE;
-        if ("earliest".equalsIgnoreCase(tag) || "index_earliest".equalsIgnoreCase(tag)) {
-            String value = element.getAttribute("value");
-            earliestFromElement = Integer.parseInt(value);
-
-        }
-        return earliestFromElement;
-    }
+public interface QueryCondition {
+    Condition condition();
 }
